core: remove redundant null checks in vm-analyzer
In processDownVms when detecting that the dbVm is null we used to check
if the dynamic data of the VM exists in the database. This is not needed
anymore since it must exist - if dbVm is null and the dynamic data is
not in the DB it means that it is an external VM and when detecting an
external VM we immediately insert its dynamic part to the DB and don't
continue to the processDownVms method in the same iteration.

Same goes to the updateRepository method.

Change-Id: I7e23388c12c0724eb4c852f1e6f142f5ba72fd3f
Signed-off-by: Arik Hadas <ahadas@redhat.com>