restapi: Make backend a request attribute
Currently the reference to the backend object is an attribute of each
resource implementation. This patch changes the API so that it is a
request attribute, managed by the "Current" object. This makes it easier
to introduce other filters that replace or enrich it as part of the HTTP
processing pipeline.

Change-Id: I18f0437cc37fd74fb07aa27a3df97c213cf20bcb
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>