webadmin: AsyncDataProvider#isCommandCompatible default
If a command is not present in the action version map, it's assumed to
be supported for any DC/cluster level, and thus shouldn't even be
checked by this method.

However, as an infrastructure, this method cannot assume it will be used
correctly, and needs to return the proper default (true) if used
improperly in this way.

Change-Id: Ifdf1d75f59f2eec353c4a7f879994ca2fddba42e
Signed-off-by: Allon Mureinik <amureini@redhat.com>