engine: Add External Network Provider to list of providers
External Network provider is a new type of external provider in oVirt.
This patch adds the backend changes needed for the the external
network provider UI changes.
Several classes in this patch are still empty
(like ExternalNetworkProviderProperties), they will be filled
with content when subsequent features for external networks will be added.

Change-Id: If3a1a60884cbbbcb140314a8b72352306934b202
Signed-off-by: mirecki <mmirecki@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1314375