core: Set image to illegal in storage while uploading.
Currently, when uploading an image- only the engine side sets this image
to locked/illegal- preventing any usage of that image.

However, the image is marked as legal in the storage. regardless to
having the actual image not synced with its engine's representation-
this could break the image's state in the engine as well when detaching
the image's containing domain, and importing it back.

When setting the image to illegal in the storage- importing it will make
the image illegal in the engine when importing again.

This patch makes the image marked as illegal in the storage while the
image is being uploaded.

Change-Id: Id3551b32d4d5f7c5a23de58687612890fcf8cd29
Signed-off-by: Amit Aviram <aaviram@redhat.com>