core: reducing race - avoid syncing the storage domain status when its in process
This patch reduce the chance of encountering a race condition, currently
we have a chance of encountring concurrent modifications to the storage
domain status

[1] - Operation on the domain
[2] - Timer that executes every few seconds GetStoragePoolInfo on the SPM host
and updates the domain data/pool metadata in the DB to match the data retrieved
from the SPM.

Those two process aren't synchronized which is problematic. The problem increased
when different statuses were added to indicate the operation performed on the domain
but the code in the timer ([2]) wasn't update.
As a first step for this patch makes [2] aware of this status which should reduce
drastically the chance for the race condition which is the condition we lived with so
far, looking forward we should look into improve that code and check the necessity of
it (as in V4 we'll no longer support any DC version with the pool metadata on storage).

Change-Id: I7080b74770e1b192ea03798f4afb7143a5f5a796
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1311121
Signed-off-by: Liron Aravot <laravot@redhat.com>