webadmin: fix memory leak in alerts / events footer
Fixed a large memory leak in alerts and events footer.
Each table has a 'dismiss' column and cell. The cells
were holding references to ModelProviders, which created
a circular reference. These tables also update every 5
seconds, so that was a ton of DOM nodes that were
detached and not GC-able.

Change-Id: Ib0ede93316863258f57619ce3e229417b4420e33
Bug-Url: https://bugzilla.redhat.com/1294678
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>