userportal, webadmin: use a custom patched bootstrap.js
Temporarily switch to a custom patched bootstrap.js.
Bootstrap 3.3.4 - 3.3.6 have a memory leak in the tooltips
that we are using extensively. It will take a while to get
that fixed upstream, so for now let's include the fix locally.

I addded a GWT-TODO so we remember to switch back to upstream
when its fixed.

Change-Id: Ie6e423738fa2905b24ae448bd870dd1b375b250a
X-Ref: https://github.com/twbs/bootstrap/issues/17973#issuecomment-202648725
Bug-Url: https://bugzilla.redhat.com/1294678
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>