engine: Remove the action version map notion
The action version map is a (now outdated) mapping between action types
(constants from VdcActionType) and the minimal cluster/DC levels
required to execute them. This mechanism has several inherit drawbacks:

* It's a parallel effort to the FeatureSupported, splitting the same
  logic to two places, making the codebase harder to follow and
  understand.
* It's checked in CommandBase's infrastructure, for every command run,
  adding a performance and complexity penalty to even to (most of the)
  commands that don't need it.
* It returns a generic message, with no option for a specific feature to
  add its custom message.

Versions 3.5 and 3.6, with their abundance of new features did not add a
thing to the action version map, reinforcing the notion that this
feature is being phased out in favor of the more modern FeatureSupported
infrastructure.

In 4.0, as the support for 3.4 (and lower) DC/clusters was removed, this
feature remains a mute point, a check that always succeeds. This patch
finishes the job and removes this feature entirely. The patch removes
the following:

1. The ActionVersionMap business entity
2. The action_version_map database table and all the stored
   procedures and DAO methods referencing it, including their tests
3. The now unused GetCommandsCompatibilityVersionsQuery that exposes
   this information to the frontend and the tooling around it

Change-Id: I97d4345363e1d47c4570f6bbf61a0e07f1052419
Signed-off-by: Allon Mureinik <amureini@redhat.com>