core: Async tasks - move steps/jobs handling to CommandBase
Currently the tasks related jobs/steps handling logic is within
CommandAsyncTask. When using CoCo to track task related commands we'll
sometimes want the child command endAction() to be executed by the parent
command, in that case the jobs/steps logic won't be executed as the related
code is in CommandAsyncTask.

This patch moves that code to the endAction() in CommandBase, so when having
a CoCo managed flow the behavior will remain the same.

Change-Id: I5836dbd3c6e3ff69fc1776471457affd9aa582f8
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1310426
Signed-off-by: Liron Aravot <laravot@redhat.com>