core: remove confusing 'stable' field from VmAnalyzer
The field called 'stable' was used to discover whether or not we
should call removeVmFromCache method in updateRepository. The logic
there is not trivial and buttom line is that this method is called
only when the status of the reported VM is down and the VM is
running on the monitored host (dbVm != null).

Therefore, moving the call to be proceedDownVm. This move should
have no side effects (it might not be necessary at all - but
need some more investigation before removing it) and it makes
the 'stable' field redundant.

Change-Id: Ib4c636d11f92b6915fec03701a37f5335ef42411
Signed-off-by: Arik Hadas <ahadas@redhat.com>