core: Stop image IO permissions before changing its status.
Currently, a request for deleting an image upload ticket is being made
AFTER the image's engine status is changed, which is problematic in
cases where the engine shuts down right after saving the image's state,
not revoking the write permissions.

This can give the user permissions to write data to a legal image if the
upload is completed, which we do not allow.

Change-Id: I4628dc3f7af404bb898a4e68dd3fbf9c2fd0be31
Signed-off-by: Amit Aviram <aaviram@redhat.com>