core: remove handleVmOnDown
The handling of report of Down VM from host that is different
than the one we expect the VM to run on in handleVmDown is
incorrect - we neither want it to remove the async command
nor to update the dynamic VM data.

The handling of report of Down VM from the host that we
expect the VM to run on was merged to proceedDownVm, rednering
handleVmOnDown redundant and therefore is removed.

Change-Id: Ie2d7b1994c3066e6964f1addab4a659d2004add4
Signed-off-by: Arik Hadas <ahadas@redhat.com>