engine: Add ability to check if a VDS is hosting the HE VM
Add a property to the VDS entity that tracks whether a given host is
hosting the hosted engine VM. The property is calculated and
populated in the VDS view and is not persisted.

This approach provides the simplest way to get this information
into the VDS entity since the Java 1.7 limitation on classes in
the common package prevents using convenience methods available
elsewhere in the code base. The change to the VDS view created
no significant performance issues when tested locally and at
scale.

Change-Id: I7e2609f59e588c6ab9279b74312b68dc5af06d5a
Bug-Url: https://bugzilla.redhat.com/1308861
Signed-off-by: Phillip Bailey <phbailey@redhat.com>