core: Pull up close to IIrsServer
JsonRpcIIrsServer already has a close() method, while IrsServerWrapper
does not, forcing IrsProxyData to use run time type recognition and
downcasting to understand how to shut its connection down.

This patch pulls up the close() method to the IIrsServer interface in
order to allow IrsProxyData to close it in a more Java-ish fashion.

As a side effect of this refactoring, the getHttpClient() can now be
removed (as it indeed was), as it's no longer needed externally. This
fact strengthens the conviction that adding the close() method was the
right thing to do, as now IrsServerWrapper no longer needs to expose its
internal implementation details beyond receiving an HttpClient instance
when its constructed.

Change-Id: Ia7335536bdcf1ec51b4e941f710f46994043172f
Signed-off-by: Allon Mureinik <amureini@redhat.com>