core: support uploading QCOW images.
Currently, uploading a QCOW image is disabled, since there were
no security verifications made on a QCOW image after upload finishes.

For QCOW images, the system should verify that the uploaded image is
actually QCOW, and verify that it does not have a backing file. This is
because having a backing file in the new image can expose undesired
resources to the user: it can point to any file in the host- including
disks of other VMs, etc..

This patch adds a VDS command that calls a VDSM volume verification
procedure, and makes the uploaded image legal only after doing this
verification- allowing a QCOW format to be chosen in the upload dialog.

Change-Id: I6bde0eb732f75afe0fef94aa79bdaab5687b76fd
Signed-off-by: Amit Aviram <aaviram@redhat.com>