core: fix import from configuration as clone
Apparently, when importing a VM as a new entity (clone) the REST-API
client does not necessarily set the ID of the VM in the parameters
(not as part of the VM object) and recent changes done in the import
command cause the command to ignore the request for cloning the VM
in this case.

Therefore, the decision whether to import a VM as clone is changed
to include the case that the import-as-new-entity flag is set and
the ID of the VM is not set in the parameters as well.

Change-Id: Ia750fda289f2a8da3deddce273cc15f51a062a44
Bug-Url: https://bugzilla.redhat.com/1322602
Signed-off-by: Arik Hadas <ahadas@redhat.com>