core: Update number of prestarted VMs after detaching a VM from Pool
Added a field to RemoveVmFromPoolParameters to update the number of
prestarted VMs in the VM Pool, if it becomes larger than number of
assigned VMs after detaching one of them. To update the number
RemoveVmFromPoolCommand takes a shared lock on the VM Pool.

This feature is activated only when a VM is detached from a Pool via UI
or REST API. It is not used (and the lock is not taken), when:

* VM Pool is removed, because prestartedVms is set to 0 at the
beginning of RemoveVmPoolCommand.
* Version of the template used for the Pool is updated, because all VMs
that are removed will be added again soon.

Change-Id: Ib864d6def677e24aa2624877e50e2ff8a8b7d1a7
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1282218
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>