core: VdcActionParametersBase - replacing shouldBeEndedByParent with enum
Currently we use a boolean (shouldBeEndedByParent) to indicate to the infra
how a child command should be ended (either when its async operations ends or
by its parent command).

In order to attempt endAction() retries in the CoCo infrastacture, this flag
should be converted to an enum. The reason is that when using a boolean flag,
in case of retrying to execute the parent endAction() it won't be able to detect
which child commands it should "collect" in CommandBase.handleChildCommands()
as their shouldBeEndedByParent flag was already updated.

This patch adds an enum called EndProcedure to replace the boolean, when the parent
"takes over" a child command this flag will be updated to indicate that.

Change-Id: I61b201dcc7742975a2d363be252ea00003adaf48
Signed-off-by: Liron Aravot <laravot@redhat.com>