core: imageupload: don't succeed if session can't be stopped.
Currently, an image uploading session succeeds regardless of succeeding
to actually stop the transfer session. This can cause having a disk ready
to be used by a VM, which is still open for external I/O to be written
into it.

This patch verifies that stopping the session succeeds before proceeding
to mark the image "Legal", therefor ready to be used.

Change-Id: I325c47f575a9fbc40a9653a7ab2178f45086ff41
Signed-off-by: Amit Aviram <aaviram@redhat.com>