core: Console related VmDynamic properties changed together
VmDynamic business entity has 3 properties identifying:
consoleCurrentUserName, consoleUserId and clientIp. Client IP address is
updated when the user connects to and disconnects from console,
"consoleUserId" is updated in SetVmTicketCommand and when VM is started.
I.e. it survives console disconnection.

Before this patch "consoleCurrentUserName" was set similarly as
"consoleUserId" but also set to `null` during user disconnection form
console. Thus after console disconnect "consoleCurrentUserName" did not
correspond to "consoleUserId"

This patch unifies behavior of "consoleCurrentUserName" and
"consoleUserId" so that they both survives console disconnect.
Frontend is altered to to behave the same way as before i.e. "Console
User" is only showed if the user is actually connected.

Change-Id: Iff69ffd09a252f24bd66981ffbde80830ef74e5b
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>