packaging: setup: asynctasks: do not start/stop engine
async tasks handling/cleanup during setup/upgrade is protected from
adding additional tasks by moving the engine to "maintenance mode".

Without this patch, we also try to start the engine when setting
maintenance mode, and stop it when we finish.

This breaks the migration flow described in the referenced bug,
where we run setup of version X where engine of version X is already
installed, but the database is from an older version. In this flow,
we do not want to start the newer engine to run against an older
database.

It's also ugly in other ways: In a normal upgrade, if the engine was
up, we needlessly try to start it. Later on, this code takes the engine
down, while we immediately try doing that again (at transaction begin).

Stop doing that. Just set maintenance mode in the database and expect
the engine to behave accordingly, if it's up.

If the engine is down, and there are pending tasks, notify the user
accordingly.

Change-Id: I4f9d31dbb6b748ac3c49ceb7d7d2b53cbd6d38df
Bug-Url: https://bugzilla.redhat.com/1318580
Signed-off-by: Yedidyah Bar David <didi@redhat.com>