core: Create vm_templates_view without other entity types
The search query used currently for Templates view looks over
vm_templates_view that contains several different entity types and then
filters templates out of the result that was returned from SQL query.
But this works badly with limited queries - the query that is limited
to return one page of results (for example, 100 items) may return less.
From Frontend point of view, the page that is not full should be the
last one.

To fix this, a new view must be created containing only template
entities. The post-filtering will not be needed and the query will
always return a full page of results for all pages except the last one.

Change-Id: I244c9a22fd3d8bd9e0b3ac2ebea717740fc030d9
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1317789
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>