core: monitoring removes jobs of down vms
Previously the callback of the live-merge command used to handle the
case where VM went down in the middle of the operation. This is wrong,
it should be done by the monitoring - when it detects that a VM goes
down it should remove all the jobs that are associated to it.

Change-Id: Ia3498e8bc79e0d0530b285f43a24bf47d4accd8f
Signed-off-by: Arik Hadas <ahadas@redhat.com>