webadmin: eliminate unnecessary (and complicating) events.
previous code fired serie of events — first it fired update change of
version field, but informing about update to some not real value. Then
real value was calculated and field was changed to that
value(multiple times). So
there was a) unnecessary events b) one event which was wrong.

after update only one event is generated and with correct value.

Motivation: following patch wants to update switch type field value
if value of version field is changed to version older than 4.0.
Because of that, we cannot tolerate any 'version updated to 3.6'
invalid event even if it's followed immediately with 'version updated
to 4.0' . Such event can be fired ONLY if version was actually changed
to 3.6.

Change-Id: I20064dfa01b068c8e43472507519d782e6c98ba3
Signed-off-by: Martin Mucha <mmucha@redhat.com>