core: can not migrate vm using REST
The regression introduced by 7d5f94d4c79ebc93cdfe08a7709a6411dda8fce7

The issue is that if the getParameters().getTargetClusterId() is null it means
that the target cluster id is not set meaning the cluster did not change.
In this case the additional permission does not have to be added.

The problem was that if the getTargetClusterId was null the Objects.equals
returned false so the additional permission has been added in a way that you
need permission to create_vm on the cluster: null which did not pass.

Change-Id: Ia0d6201c3d095beebffb47a6fb6beab5e5684b6f
Bug-Url: https://bugzilla.redhat.com/1328011
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>