engine: Stop checking if hosts support live snapshot
Old VDSM versions on EL servers required qemu-kvm, which is a stripped
down upstream flavor of qemu that does not support live snapshots. This
created a situation where even though a host had a sufficiently high
VDSM version to be part of a cluster that theoretically allowed live
snapshots (i.e., 3.1 or above), the host itself may not have supported
this feature, leading to the need to check for it explicitly, regardless
of the cluster version.

Modern VDSM versions (specifically, 4.17 or above, which is the minimal
requirement to connect to the current engine) require qemu-kvm-[rh]ev in
their spec file, promising the ability to perform live snapshots.

This commit removes the liveSnapshotSupported property from VdsDynamic,
VDS, the relevant database objects and any [redundant ]validations
around it.

Note that this property is visible in the REST API, although it's a
read-only property which cannot be updated or manipulated.
As the REST API model currently still specifies this property, the
current patch does not remove it, but instead stubs it with a hardcoded
`true` value. A subsequent patch will require a newer
ovirt-engine-api-model version (which is not yet released at the time
this patch was authored) to remove it from the API too.

Change-Id: I9f1bcbd898885b999a57184eba8bc65c9c993900
Signed-off-by: Allon Mureinik <amureini@redhat.com>