engine: Stop checking if hosts support live merge
oVirt Engine 3.5.1 introduced the live merge feature in the 3.5 DC
level. However, this DC version allowed RHEL 6.y and RHEL 7.0 hosts with
[relatively] old qemu and libvirt versions which did not support it,
forcing the engine to explicitly check if the host running a VM a
snapshot should be merged on supports the feature.

Since commit 6eecdd7 only DCs of level 3.6 or higher are supported,
meaning that the underlying platform is EL 7.1 or Fedora 22 (at least).
The qemu and libvirt versions there support live merging snapshots,
making this check redundant.

Removing this check not only cleans up the code considerably, but also
reduces execution time in a noticeable way by removing additional
database flows from various places in the flow.

Change-Id: I227efa20d9d7c26c15d536c378734563c87887dc
Signed-off-by: Allon Mureinik <amureini@redhat.com>