restapi: Remove Host#liveSnapshotSupport from V4 API
The previous patch removed the liveSnapshotSupport property from the
backend VDS object and the related classes, but did not change the API
(other than hollowing it out).

This patch completes the job by consuming the updated
ovirt-engine-api-model artifact that removes this property from the new
V4 API.

In order to preserve backwards compatibility, the V3 API will continue
to report a `true` value, which is set in a hard-coded manner. Note that
even in the previous version this is a read-only property which could
not be updated (neither via the API nor by any other action via the
engine - the only way to affect this property is, theoretically, to
install/remove/update the qemu-kvm-[rhev] package on the host, but see
below). Additionally, any host that can participate in a 3.6 or newer
cluster will have this property as `true` anyway, as explained in the
previous patch.

Change-Id: I160d7ed8742128fdaac88bdc2228f4739139c5f5
Signed-off-by: Allon Mureinik <amureini@redhat.com>