core: Make disk list of VM editable on HE update
Disks are never updated by the UpdateVmCommand but the editable field
checks are checking the value if it changed. The UI will not send the
disk list with the VM when it updates a VM but the REST API does.

Mark the diskList field as editable to make updating the Hosted Engine
always work.

Change-Id: I1ed00104d929b71de0ceb6d6cce9482946a8f5ab
Bug-Url: https://bugzilla.redhat.com/1326810
Signed-off-by: Roman Mohr <rmohr@redhat.com>