core: audit console's client ip change in vm-analyzer
There is no reason not to audit a change in console's client ip in
VmAnalyzer. We do not call bll command so no need to put it in
VdsEventListener. We do not do anything in a batch for all VMs
reported on the same host so no need to do it in VmsMonitoring.
Thus, moving it to VmAnalyzer.

Change-Id: I5a9aafa9820582f11c6b70c1fb7e048c030f5591
Signed-off-by: Arik Hadas <ahadas@redhat.com>