engine: add openstackNetworkProviderId to VdsStatic
Adding openstackNetworkProviderId to VdsStatic.
This is required to distinguish hosts which have the
neutron agent configured during host installation.
In this case the neutron agent has the property
binding:host_id set to a custom value, and this
needs to be reflected during openstack port allocation.
In case the agent is installed manually, the property
value must not be customized.

Change-Id: Idd70f7ab57568379862e969abd732eaa4634c595
Signed-off-by: mirecki <mmirecki@redhat.com>