engine: Moving network provider id from paramters to paramters.vdsstaticdata
The openstack network provider is moved from parameters to
parameters.vdsstaticdata. This will alow to persist
the openstack network provider as part of VdsStatic data.
This is required to distinguish hosts which have the
neutron agent configured during host installation.
In this case the neutron agent has the property
binding:host_id set to a custom value, and this
needs to be reflected during openstack port allocation.
In case the agent is installed manually, the property
value must not be customized.

Change-Id: I71c6ac97486a210d80da6e5b6e23386b32e538ff
Signed-off-by: mirecki <mmirecki@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1318543