engine: Set the vfsType glusterfs when importing HE domain
In HE flow, when importing a gluster domain we must set explicitly the
domain type otherwise the mount will fail.

This problem discovered when non HE hosts where connecting to that
domain. HE hosts have the right mount options in a configuration file
(they are connecting before there is an engine live)

Change-Id: Ic2383047d5033818ae2c2227e9fd6e756db10eaa
Bug-Url: https://bugzilla.redhat.com/1317699
Signed-off-by: Roy Golan <rgolan@redhat.com>