core: Live Merge: Improve Live Merge Recovery Mechanism
When sending live merge command, we may fail to get vdsm's response
due to network error. In this case, the parent of the volume that
should have been deleted (the volume that the data is merged in to)
will be marked as ILLEGAL. However, a merge job may have started, and
even possibly completed, and we want to allow the user to try live
merge again in order to remove the snapshot. When the user tries live
merge again, if the volume is already deleted from the chain or if
there is a running merge job, vdsm returs 'imageErr' or 'mergeErr',
respectively. In this patch, engine catches those errors and
continues the merge flow as if it succeeded. If merge succeeded the
volume is removed from database. Otherwise, volume stays in ILLEGAL
state.

Change-Id: Ic54ae8576410afbdd095fdae1fb298f2262218f5
Bug-Url: https://bugzilla.redhat.com/1301083
Bug-Url: https://bugzilla.redhat.com/1304761
Bug-Url: https://bugzilla.redhat.com/1314082
Signed-off-by: Ala Hino <ahino@redhat.com>