webadmin: Enable tooltip on greyed out LUNs
When editing an ISCSI/FC storage domain, if a LUN is
already in use it will be greyed out.
If the user hovers over the LUN a tooltip appears
with the reason why it is disabled.

In this patch, the tooltip is added to the whole row,
not only the checkbox cell.
It will always appear on the left of the row.

Bug-Url: https://bugzilla.redhat.com/1329906

Change-Id: I4ea714768c4e427cf82a8b062fd49dcaf127335b
Signed-off-by: Fred Rolland <frolland@redhat.com>
Signed-off-by: Daniel Erez <derez@redhat.com>