core: Fix permissions for profiles created for new domain
The AddDiskProfileCommand takes care of the permissions for newly
added disk profiles and gives the users in Everyone group the
DiskProfileUser role.
AddStorageDomain command added the created disk profile directly
to the DB without calling the command and so this role was not
assigned.
The fix was to call AddDiskProfileCommand directly from
AddStorageDomain and also from AddStorageDomainToPoolCommand which
creates a disk profile in case one doesn't exist for the storage domain.

Change-Id: Ia6dd5483489b65285cbd5905865c07b9446f6459
Bug-Url: https://bugzilla.redhat.com/1209505
Signed-off-by: Jenny Tokar <jtokar@redhat.com>