core: Remove Entities#getIds
In Java 8, with a sensible steaming API, there's no reason to have a
method like Entities#getIds.

Removing this method not only improves the code by removing spam, in
some cases, improves performance, where a Set is used instead of a List,
improving the complexity of the contains operation to O(1) instead of
O(n).

Change-Id: Ia6008fd6701322a62f1e734b8fda925fda6e2dd6
Signed-off-by: Allon Mureinik <amureini@redhat.com>