core: Remove Entities#objectNames
In Java 8, with a sensible steaming API, there's no reason to have a
method like Entities#objectNames.

Removing this method not only improves the code by removing spam, in
most cases, where the names are just joined to a comma delimited string,
it also improves performance - instead of traversing two collections
(one of Namables and then one of Strings), it can be streamed just once.

Change-Id: Id47577a4bb671ee5baca16aac63648c9c7e9c41a
Signed-off-by: Allon Mureinik <amureini@redhat.com>