core: Enable symlink following in all web applications
The web container included in WildFly doesn't follow symlinks in
exploded web applications by default. We need that for the branding
mechanism, as it is implemented with a symlink to the "branding.jar" in
the "WEB-INF/lib" directory of the applications. In order to support
that this patch adds to all the web applications a WildFly specific
deployment descriptor (the "jboss-web.xml" file) that enables following
of symlinks.

Change-Id: Ieff80a8eac50e25c1d30e3e17c6d01cd5bc70e89
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>