webadmin: remove 'unsupported browser' warning
Commit db95728 introduced a popup message in the webadmin
header that warned people when their browser was
"unsupported." However, we received many complaints that
this message was simply confusing.

Since we haven't have a browser compatibility issue in a
very long time, we decided it was best to just remove the
message.

It's safe to say that we can support "any modern
browser." I'll add something to that effect on the oVirt
website.

Change-Id: I2094c7640d4bda243e7b524f463962337398ab7f
Bug-Url: https://bugzilla.redhat.com/1326511
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>