restpi: BackendStorageDomainResourceTest#testRemoveWithHostId
This method was declared as a JUnit 3 style test (public void method
with no arguments and a name that starts with the word "test"), but the
class is run under a JUnit 4 runner, so this test is never run.

This patch adds the required @Test annotation to make it a real test.

Change-Id: Id7621976443322c43cebf8e3d50f11600081ca57
Signed-off-by: Allon Mureinik <amureini@redhat.com>