Frontend: add VM name to console title of noVNC,SPICE HTML5
In case of opening a console to VM using either noVNC client or
SPICE HTML5 clients (both consoles running inside the browser) -
the VM name wasn't displayed in title.

before this fix:
for noVNC - the console title was "noVNC".
for SPICE HTML5 - the console title was "Spice Javascript Client".

This patch added the VM name to the title for both protocols.

Bug-Url: https://bugzilla.redhat.com/1313295
Change-Id: If3897b7932e2fa5bd0567c091c47751c32e07675
Signed-off-by: Sharon Gratch <sgratch@redhat.com>