webadmin: Synched Cpu Pinning label and textbox
"Cpu Pinning Topology" label wasn't in sync with the availability of
it's textbox. It was constantly greyed out(disabled) even when the
textbox was enabled.

This patch introduced synching with between the textbox and the label.

Due to the blue tooltip between the label and the textbox it's impossible
to use the combined label and textbox object. So, This patch is using
listeners inorder to overcome this problem.

Change-Id: I007814512506170699050a4cb82abfea2278c4d4
Bug-Url: https://bugzilla.redhat.com/1226767
Signed-off-by: Tomer Saban <tsaban@redhat.com>