core: Update getTotalMemorySizeInBytes
VM.getTotalMemorySizeInBytes apparently uses outdated value for
computing VM's video RAM.  This patch retrieves video RAM size from the
corresponding helper.

We must move the method to modules/bll so that we are able to use the
video RAM size helper.  We also renamed the method to be more
distinctive from the recently introduced getTotalRequiredMemoryInMb
method.

Change-Id: I2f0fd1ae3f1894b51ee8983725c8763dafbb0503
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>