webadmin: Adding missing aggregate link tooltip to setup networks panel
Bonds in mode 4 (link aggregation) have an aggregate assigned when working properly.
This mac is reported by vdsm. If this mac is missing, the condition must be reported
to the user.
This patch adds an alert icon to the host network interfaces subtab,
which notifies the user that such a contition has occured.

Change-Id: I73fc50bd59abf09f8372cf459126dfedb0b159e7
Bug-Url: https://bugzilla.redhat.com/1281666
Signed-off-by: mirecki <mmirecki@redhat.com>