engine: add network_filter_id column to vnic_profiles
The column will be used to configure each vNIC's network filter.

If 'Vdc_Options.EnableMACAntiSpoofingFilterRules' is 'true',
'vdsm-no-mac-spoofing' filter will be set on all the vNic profiles,
otherwise NULL will be set.

In case the vNIC profile points to NULL, no network filter will be
configured for that specific vNIC.

It is important to mention, that libvirt's backward compatibility
version support is assumed, meaning that libvirt will keep supporting
all the filters introduced in older versions.  Otherwise the output
for the following scenario is undefined: network N defined in data
center with compatibility version V1 is being assigned to cluster with
compatibility version V2 where V2 > V1, as the network filter of
version V1 may not be supported in hosts of version V2.

The reason for allowing this tightly coupling in libvirt is product
commonality, as other features depending on libvirt assuming backward
compatibility.

Change-Id: I1a7cce4b9b377a4eaf309ad63e87832489b25a16
Signed-off-by: Eliraz Levi <elevi@redhat.com>
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>