restapi: Fix V3 snapshot action links
In version 4 of the API all the URL segments are generated concatenating
the words of the relevant name without separators. But in version 3 some
URL segments didn't follow this rule, in particular the URL segments
corresponding to snapshot actions used an underscore to separate words:

  preview_snapshot
  commit_snapshot
  undo_snapshot

To preserve backwards compatibility this patch changes the V3 adapter so
that it fixes the links to these actions, adding the required
underscore.

Change-Id: Idf1712f58dc29d6a91335f2c2f228e9fe9f746aa
Bug-Url: https://bugzilla.redhat.com/1332986
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>