engine: change VmInfoBuilder network filter logic
The feature of Network filter enhancing the user ability to customize the
vNIC's network filter type beside 'vdsm-no-mac-spoofing' or configurating
no filter at all.

Please note that vdsm will configure the mentioned 'filter' property's
value as the vNIC's network filter.
In case no network filter is desirable, the filter property should not
be mentioned at all.

The patch is modifing the engine's logic which used to un\mention the
'filter' property according to the value of
'EnableMACAntiSpoofingFilterRules' saved in vdc_options table.

The option_name of 'EnableMACAntiSpoofingFilterRules' and the
corresponding backend's representation is kept for backward
compatibility support, specifically for Rest API scripting for determining
the default network filter behavior.

Please note that the use of NetworkFilterDao was without DI as it
required a major refactoring of VmInfoBuilder and it should be done in a
dedicated patch.

Change-Id: Ie0cb0604e708adbc09dc4742c47464131357fe32
Signed-off-by: Eliraz Levi <elevi@redhat.com>
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>