core: handle failed internal migrations properly
InternalMigrateVmCommand used to always do setSucceeded(true) no matter
if the migration was triggered successfully or not (MigrateVmCommand does
not throw exception in the execute phase in case of failures).

This patch changes InternalMigrateVmCommand to set the result properly
and changes MaintenanceVdsCommand to check whether the migration has
failed or not properly, by checking its getSucceeded().

Change-Id: I4289365a1aceb02adea5029beac1547e024ebca4
Bug-Url: https://bugzilla.redhat.com/1332039
Signed-off-by: Arik Hadas <ahadas@redhat.com>